backend/requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Gmail API dependencies
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0

# LLM integration
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0